Moved fakes to a separate file usable by other tests
